Created friend_request message, fixed some bugs, added more onion helper functions

+++ include/db_options.h
@@ -20,15 +20,15 @@ int db_options_get_int(sqlite3 *db, const char *key);
 void db_options_set_int(sqlite3 *db, const char *key, int value);
 
 // Fetch the option of binary object (BLOB) type, returns number of bytes written
-int db_options_get_bin(sqlite3 *db, const char *key, void *value, size_t value_len);
+int db_options_get_bin(sqlite3 *db, const char *key, void *value, int value_len);
 
 // Set value of binary type option
-void db_options_set_bin(sqlite3 *db, const char *key, const void *value, size_t value_len);
+void db_options_set_bin(sqlite3 *db, const char *key, const void *value, int value_len);
 
 // Fetch the option of text type
-int db_options_get_text(sqlite3 *db, const char *key, char *value, size_t value_len);
+int db_options_get_text(sqlite3 *db, const char *key, char *value, int value_len);
 
 // Set value of text type option, value_len can be -1 for entire null terminated string
-void db_options_set_text(sqlite3 *db, const char *key, const char *value, size_t value_len);
+void db_options_set_text(sqlite3 *db, const char *key, const char *value, int value_len);
 
 #endif
